+++ packages/preset-mini/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './colors'
 export * from './mappings'
 export * from './handlers'
 export * from './variants'
